+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [v0.3.0] - Work In Progress
+### Changed
+- Renamed `applications import get_application_from_module` as `get_applications_using_module`
+- Shortened import paths: no more `$resource.$resource`
+
 ### Added
+- `black` code formatter
 - New commands to generate & validate local files based on templates:
 
     hesperides local-generate-files tests/descriptor.json